requirements.txt
Cython==0.23.5
thriftpy==0.3.7
SQLAlchemy==1.0.12
MySQL-python==1.2.5
arrow==0.7.0
six==1.10.0
PyYAML==3.11
requests==2.9.1
hiredis==0.2.0
redis==2.10.5
PySimpleSOAP==1.16
apscheduler==3.1.0
